+++ hamming/src/hamming.h
@@ -1,6 +1,6 @@
 #ifndef HAMMING_H
 #define HAMMING_H
 
-int compute(const char *lhs, const char *rhs);
+long long compute(const char *lhs, const char *rhs);
 
 #endif
